% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_school_pop.R
\name{build_school_pop}
\alias{build_school_pop}
\title{Helper function to identify core school population}
\usage{
build_school_pop(year = "2014")
}
\arguments{
\item{year}{character. The survey year of interest.  Default is 2014}
}
\value{
a dataframe of basic school information, which is a subset of all schools in the IPEDS universe
}
\description{
A function to filter down school data to generate a school population for further analyses.
}
\examples{
\dontrun{
school_pop  <- build_school_pop()
}
}

