% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charges_ay.R
\name{charges_ay}
\alias{charges_ay}
\title{Student Charges for Academic Year programs.}
\format{Data frame with columns}
\source{
http://nces.ed.gov/ipeds/datacenter/DataFiles.aspx
}
\usage{
charges_ay
}
\description{
For more information, download a data dictionary from the IPEDS website.
}
\details{
Survey years 2002 - 2014.
}

