% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk of various codes a US-based Higher Education institution may have}
\format{Data frame with columns}
\source{
Original data from: http://www.pesc.org/interior.php?page_id=145
}
\usage{
crosswalk
}
\description{
These codes allow analysts to join datasets from the College Board, ACT, and various Department of Ed codes
}

