% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efc.R
\docType{data}
\name{efc}
\alias{efc}
\title{Fall Enrollment: Residence and Migration of first-time Freshmen}
\format{Data frame with columns}
\source{
http://nces.ed.gov/ipeds/datacenter/DataFiles.aspx
}
\usage{
efc
}
\description{
For more information, download a data dictionary from the IPEDS website.
}
\details{
Survey years 2002 - 2014.
}
\examples{
  efc
}
\keyword{datasets}

