% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efcp.R
\name{efcp}
\alias{efcp}
\title{Fall Enrollment: Major Field of Study by r/e, gender, attendance status, and level}
\format{Data frame with columns}
\source{
http://nces.ed.gov/ipeds/datacenter/DataFiles.aspx
}
\usage{
efcp
}
\description{
For more information, download a data dictionary from the IPEDS website.
}
\details{
Survey years 2002 - 2014, every other year (even years)
}

