% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipeds.R
\name{get_ipeds}
\alias{get_ipeds}
\title{Download IPEDS files}
\usage{
get_ipeds(
  year = 2022:2023,
  survey,
  title = "directory",
  file_name,
  local_dir = NA,
  revised_files = TRUE
)
}
\description{
Download selected IPEDS files and place into a tibble
}
