% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.R
\docType{data}
\name{hd}
\alias{hd}
\title{Institutional Directory Information (HD)}
\format{Data frame with columns}
\source{
http://nces.ed.gov/ipeds/datacenter/DataFiles.aspx
}
\usage{
hd
}
\description{
For more information, download a data dictionary from the IPEDS website.
}
\details{
Survey years 2002 - 2014.
}
\examples{
  hd
}
\keyword{datasets}

