% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\docType{data}
\name{ic}
\alias{ic}
\title{Institutional Characteristics and Admissions Data from IPEDS}
\format{Data frame with columns}
\source{
http://nces.ed.gov/ipeds/datacenter/DataFiles.aspx
}
\usage{
ic
}
\description{
For more information, download a data dictionary from the IPEDS website.
}
\details{
Survey years 2002 - 2014.

Note:  In 2014, the data in the IC survey files was broken out into IC and ADM.  These files are combined.
}
\examples{
  ic
}
\keyword{datasets}

