% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripeds.R
\name{ipeds.fetch}
\alias{ipeds.fetch}
\title{IPEDS fetch}
\usage{
ipeds.fetch(year)
}
\arguments{
\item{year}{}
}
\value{
list, containing data.frames of all the data for the provided year
}
\description{
For the year
}
