% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ipeds_dict}
\alias{ipeds_dict}
\title{Search data dictionary.}
\usage{
ipeds_dict(
  search_string,
  search_col = c("all", "description", "varname", "filename"),
  exact_match = FALSE,
  limit = 10,
  confirm = FALSE,
  return_dict = FALSE,
  print_off = FALSE
)
}
\arguments{
\item{search_string}{Character string for search. Can use regular expression
for search. Must escape special characters, \code{. \ | ( ) [ \{ ^ $ * +
  ?}, with a doublebackslash \code{\\\\}.}

\item{search_col}{Column to search. The default is to search all columns.
Other options include: "varname", "description", "filename".}

\item{exact_match}{Set to TRUE if you only want exact search string matches.
Note that setting exact_match to TRUE may not work as expected if search string
includes regular expressions.}

\item{limit}{Only the first 10 dictionary items are returned by default.
Increase to return more values. Set to \code{Inf} to return all items
matched in search'}

\item{confirm}{Use to confirm status of variable name in dictionary. Returns
\code{TRUE} or \code{FALSE}.}

\item{return_dict}{Return a tibble of the subset data dictionary.}

\item{print_off}{Do not print to console; useful if you only want to return a
tibble of dictionary values.}
}
\description{
This function is used to search the IPEDS data dictionary.
}
\section{Interactive dictionary}{


When using \code{ipeds_dict()} interactively (the default), formatted output will
be returned to the console.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{> ipeds_dict("efydesom")

=====================================================================
VARIABLE: efydesom
=====================================================================

:::::::::::::::::::          DESCRIPTION          :::::::::::::::::::

Students enrolled in some but not all distance education courses

:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::

../FILES

 |__ EFFY2020_DIST*
 |__ EFFY2021_DIST*
 |__ EFFY2022_DIST*
 |__ EFFY2023_DIST*

 * Denotes a long file in which institutions may have more than one
 record (UNITID values repeated across multiple rows).

=====================================================================
Printed information for 1 of out 1 variables.
}\if{html}{\out{</div>}}

Output includes:
\itemize{
\item Variable name
\item Each unique description found across dictionary files provided by NCES
\item All complete data files in which the variable with that description is found
\item A flag (*) if necessary to denote the data file is a long file
\item A note to indicate if more results are available than the limit are available
}
}

\examples{
\dontrun{
## simple search for "state" in any part of the dictionary
ipeds_dict("state")

## variable names starting with "st"
ipeds_dict("^st", search_col = "varname")

## return full dictionary (only recommended if not printing and
## storing in object)
df <- ipeds_dict(".", limit = Inf, print_off = TRUE, return_dict = TRUE)
}
}
