% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_table.R
\name{ipeds_file_table}
\alias{ipeds_file_table}
\title{All currently available complete data files}
\usage{
ipeds_file_table(redownload_table = FALSE)
}
\arguments{
\item{redownload_table}{Reconnect to URL to download table of files (FALSE)}
}
\description{
Return a table with currently available data files. The first request in a
session will scrape table from NCES website. Subsequent function calls in the
same session will return saved table unless user requests to redownload the
table.
}
