% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_filter.R
\name{ipeds_filter}
\alias{ipeds_filter}
\title{Filter IPEDS data by variable values.}
\usage{
ipeds_filter(ipedscall, ...)
}
\arguments{
\item{ipedscall}{Current list of parameters carried forward from prior
functions in the chain (ignore)}

\item{...}{Expressions to evaluate}
}
\description{
This function is used to filter downloaded IPEDS data.
}
\examples{
\dontrun{
}
}
