% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ipeds_get}
\alias{ipeds_get}
\title{Get IPEDS data.}
\usage{
ipeds_get(ipedscall, bind = TRUE, join = TRUE)
}
\arguments{
\item{ipedscall}{Current list of parameters carried forward from prior
functions in the chain (ignore)}

\item{bind}{Row bind all same name survey files (e.g., HD2022 and HD2023)}

\item{join}{Join different name survey files by UNITID and year. If \code{bind = FALSE}, then \code{join} will be set to \code{FALSE} and the function argument
ignored.}
}
\value{
Depending on argument combination, the chain will return one of the
following objects:
\enumerate{
\item \verb{bind = FALSE, join = FALSE}: A list of files with no further
processing (each unique complete data file required returned as a list item).
\item \verb{bind = TRUE, join = FALSE}: A list of files in which like files
(e.g., HD*, IC*) are row bound together but unjoined to unlike files
\item \verb{bind = TRUE, join = TRUE}: A data frame in which like files
are bound and all are joined
}

NOTE: The more complicated the data call (many selected variables, many
selected years, more complex filter), the longer the data request may take,
particularly if downloading files, and the greater the likelihood of
unexpected behavior with the join. Users may wish to break up large complex
requests into multiple smaller requests or elect to return a list of unbound
/ unjoined data frames they can manipulate directly.
}
\description{
Retrieve IPEDS data by establishing which data files are required for
filtering (if included) and those required for selected variables.
}
\examples{
\dontrun{
# default: bind = TRUE, join = TRUE
ipeds_get()

# bind only
ipeds_get(join = FALSE)

# non-bind, non-join
ipeds_get(bind = FALSE, join = FALSE)
ipeds_get(bind = FALSE) # join will be set to FALSE by default
}

}
