% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ipeds_temp_to_disk}
\alias{ipeds_temp_to_disk}
\title{Save downloaded IPEDS files in temporary directory to disk}
\usage{
ipeds_temp_to_disk(
  to_dir,
  overwrite = FALSE,
  remove_from_tempdir = FALSE,
  create_dir = TRUE
)
}
\arguments{
\item{to_dir}{Directory path for saving zip files (will be created if it does
not exist unless \code{create_dir} is set to \code{FALSE}).}

\item{overwrite}{Overwrite files in local directory with those in temporary
directory.}

\item{remove_from_tempdir}{Remove files from temporary directory after saving
in local directory (they otherwise will be removed at the end of the R
session by default).}

\item{create_dir}{Recursively create directory path specified by user if it
does not exist. Default is \code{TRUE}.}
}
\description{
Save IPEDS data files (zip) stored in temporary directory to disk for future
use. By default, the function saves a copy of the files in the local
directory while leaving a copy in the temporary directory. In addition the
function will create subdirectories as necessary to save the files.
}
