% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_disk.R
\name{ipeds_tmp_to_disk}
\alias{ipeds_tmp_to_disk}
\title{Save downloaded IPEDS files in temporary directory to disk.}
\usage{
ipeds_tmp_to_disk(
  to_dir,
  overwrite_existing = FALSE,
  remove_from_tempdir = FALSE,
  create_directory = TRUE
)
}
\arguments{
\item{to_dir}{Directory path to copy}

\item{overwrite_existing}{Overwrite files on local directory with those in
temporary directory}

\item{remove_from_tempdir}{Remove files from temporary directory after saving
in new location}

\item{create_directory}{Recursively create directory path specified by user
if it does not exist}
}
\description{
Save IPEDS zip files currently stored in temporary directory to disk for
future usage. By default, the function makes a copy of the files in the new
directory, leaving a copy in the temporary directory. By default, the
function will create directories as necessary to save the files.
}
