% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{ipeds_year}
\alias{ipeds_year}
\title{Select scorecard data year.}
\usage{
ipeds_year(ipedscall, year)
}
\arguments{
\item{ipedscall}{Current list of parameters carried forward from prior
functions in the chain (ignore)}

\item{year}{Four-digit year or string \code{latest} for latest data.}
}
\description{
This function is used to select the year of the data.
}
\section{Important notes}{

\enumerate{
\item Not all variables have a year option.
\item The year selected is not necessarily the year the data were produced.
It may be the year the data were collected. For data collected over split
years (fall to spring), it is likely the year represents the fall data (\emph{e.g.,}
2011 for 2011/2012 data).
}
}

\examples{
\dontrun{
ipeds_year(2012)
ipeds_year(2010:2013)
}
}
