% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripeds-package.R
\docType{package}
\name{ripeds-package}
\alias{ripeds}
\alias{ripeds-package}
\title{ripeds: Download and Manipulate Integrated Postsecondary Education Data System (IPEDS) Data}
\description{
Download and manipulate Integrated Postsecondary Education Data System (IPEDS) complete data files using a series of piped functions to facilitate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://potential-adventure-6knn6m6.pages.github.io/}
}

}
\author{
\strong{Maintainer}: Benjamin Skinner \email{bskinner@neh.gov} (\href{https://orcid.org/0000-0002-0337-7415}{ORCID})

}
\keyword{internal}
