% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa.R
\docType{data}
\name{sfa}
\alias{sfa}
\title{Student Financial Aid and Net Price information}
\format{Data frame with columns}
\source{
http://nces.ed.gov/ipeds/datacenter/DataFiles.aspx
}
\usage{
sfa
}
\description{
For more information, download a data dictionary from the IPEDS website.
}
\details{
Aid years 2002-03 to 2013-14.
}
\examples{
  sfa
}
\keyword{datasets}

