% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site1_network.R
\name{site1_network}
\alias{site1_network}
\title{A crawl of the network of data from a College Search site.}
\format{2 data frames.  Links (edgelist, with rank order shown), and majors (schools and majors offered)}
\source{
myplan.com
}
\usage{
site1_network
}
\description{
The crawl occurred on April 3rd, 2016.
}
\details{
The data connects schools (Other schools to Consider) and are based on UNITID,  Majors are represented by CIP Codes.
}

