% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site2_network.R
\name{site2_network}
\alias{site2_network}
\title{A crawl of the network of data from a College Search site.}
\format{2 data frames.  edges (edgelist, with rank order shown), and mission statements (schools and mission statement text)}
\source{
cappex.com
}
\usage{
site2_network
}
\description{
The crawl occurred on April 3rd, 2016.
}
\details{
The data connects schools (Similar) and are based on UNITID,  The school's mission statement, when available (via the site) are also included
}

