% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipeds.R
\name{get_ipeds}
\alias{get_ipeds}
\title{Download IPEDS files}
\usage{
get_ipeds(
  year,
  survey,
  title,
  file,
  local_dir = NA,
  revised_files = TRUE,
  lower_names = TRUE,
  filter_bool = c("and", "or"),
  bind = TRUE,
  join = FALSE
)
}
\description{
Download selected IPEDS files and place into a tibble
}
