% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripeds.R
\name{ipeds.fetch}
\alias{ipeds.fetch}
\title{IPEDS fetch}
\usage{
ipeds.fetch(year)
}
\arguments{
\item{year}{integer, the 4 digit year for which data will be downloaded}
}
\value{
list, containing data.frames of the data for the provided year
}
\description{
Downloads and reads data from the NCES IPEDS for the year provided.
}
