% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ipeds_dict}
\alias{ipeds_dict}
\title{Search data dictionary.}
\usage{
ipeds_dict(
  search_string,
  search_col = c("all", "description", "varname", "filename"),
  ignore_case = TRUE,
  limit = 10,
  confirm = FALSE,
  return_dict = FALSE,
  print_off = FALSE
)
}
\arguments{
\item{search_string}{Character string for search. Can use regular expression
for search. Must escape special characters, \code{. \ | ( ) [ \{ ^ $ * +
  ?}, with a doublebackslash \code{\\\\}.}

\item{search_col}{Column to search. The default is to search all columns.
Other options include: "varname", "description", "filename".}

\item{ignore_case}{Search is case insensitive by default. Change to
\code{FALSE} to restrict search to exact case matches.}

\item{limit}{Only the first 10 dictionary items are returned by default.
Increase to return more values. Set to \code{Inf} to return all items
matched in search'}

\item{confirm}{Use to confirm status of variable name in dictionary. Returns
\code{TRUE} or \code{FALSE}.}

\item{return_dict}{Return a tibble of the subset data dictionary.}

\item{print_off}{Do not print to console; useful if you only want to return a
tibble of dictionary values.}
}
\description{
This function is used to search the IPEDS data dictionary.
}
\examples{
## simple search for "state" in any part of the dictionary
ipeds_dict("state")

## variable names starting with "st"
ipeds_dict("^st", search_col = "varname")

## return full dictionary (only recommended if not printing and
## storing in object)
df <- ipeds_dict(".", limit = Inf, print_off = TRUE, return_df = TRUE)
}
