% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_to_disk.R
\name{ipeds_download_to_disk}
\alias{ipeds_download_to_disk}
\title{Download IPEDS files directly to disk.}
\usage{
ipeds_download_to_disk(
  files,
  to_dir,
  overwrite_existing = FALSE,
  create_directory = TRUE
)
}
\arguments{
\item{files}{Vector of files (without ending) to download (e.g., HD2023)}

\item{to_dir}{Directory path to copy}

\item{overwrite_existing}{Overwrite files on local directory with those in
temporary directory}

\item{create_directory}{Recursively create directory path specified by user
if it does not exist}
}
\description{
Directly download and save IPEDS zip files to disk for future usage. By
default, the function will create directories as necessary to save the files.
}
