% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ipeds_download_to_disk}
\alias{ipeds_download_to_disk}
\title{Download IPEDS files directly to disk}
\usage{
ipeds_download_to_disk(
  to_dir,
  files = NULL,
  use_ipeds_dict = NULL,
  type = c("data", "dictionary"),
  overwrite = FALSE,
  create_dir = TRUE
)
}
\arguments{
\item{to_dir}{Directory path for saving zip files (will be created if it does
not exist unless \code{create_dir} is set to \code{FALSE}).}

\item{files}{Vector of survey file names without ending to download (e.g.,
HD2023).}

\item{use_ipeds_dict}{Output from \code{\link[=ipeds_dict]{ipeds_dict()}} in which \code{return_dict = TRUE};
if argument isn't missing, it will be chosen over input to files argument
if also included.}

\item{type}{Download the data file (default) or associated dictionary file.}

\item{overwrite}{Overwrite files on local directory with those in temporary
directory. Default is \code{FALSE} (keep already existing files).}

\item{create_dir}{Recursively create directory path specified by user if it
does not exist. Default is \code{TRUE}.}
}
\description{
Directly download and save IPEDS zip files to disk for future usage. By
default, the function will create subdirectories as necessary in which to
save the files.
}
\examples{
\dontrun{
ipeds_download_to_disk(".", c("HD2020", "HD2021"))
ipeds_download_to_disk(".", c("HD2020", "HD2021"), overwrite = TRUE)

files <- c("HD2020", "HD2021")
ipeds_download_to_disk(".", files)

# using return from ipeds_file_table() to down all 2020 files
dict <- ipeds_file_table()
ipeds_download_to_disk(".", ipeds_dict = dict[dict$year == 2020,])

# download associated dictionary files
ipeds_download_to_disk(".", c("HD2020", "HD2021"), type = "dictionary")
}

}
