% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{ipeds_file_table}
\alias{ipeds_file_table}
\title{Return a table of all currently available complete data files}
\usage{
ipeds_file_table(redownload = FALSE, return_table = TRUE)
}
\arguments{
\item{redownload}{Re-scrape NCES website to generate table of files. Defaults
to \code{FALSE}.}

\item{return_table}{Return table (Default = \code{TRUE}).}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{year}: Survey year assigned to the file by NCES
\item \code{survey}: Survey file category
\item \code{title}: Survey file title
\item \code{file}: Survey file name
}
}
\description{
Return a table with currently available IPEDS complete data files (see
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}). Unless called by
the user on its own, the first \code{\link[=ipeds_init]{ipeds_init()}} call in a session will call
\code{\link[=ipeds_file_table]{ipeds_file_table()}} to scrape the table from NCES website and store the
results in a temporary directory. Subsequent data requests during the session
will return the table in memory unless the user calls requests to redownload the
table.
}
\examples{
\dontrun{
# First call in a session will scrape NCES website and return
ipeds_file_table()

# Second call in the session will return table in memory
ipeds_file_table()

# Will re-scrape NCES website
ipeds_file_table(redownload = TRUE)
}

}
