% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{ipeds_filter}
\alias{ipeds_filter}
\title{Filter IPEDS data by variable values.}
\usage{
ipeds_filter(ipedscall, ...)
}
\arguments{
\item{ipedscall}{Current list of parameters carried forward from prior
functions in the chain (ignore)}

\item{...}{Expressions to evaluate}
}
\description{
This function is used to filter downloaded IPEDS data.
}
\examples{
\dontrun{
# using non-standard evaluation, you can separate filters by a comma, which
# will serve as an `&` (AND) boolean
ipeds_filter(stabbr == "KY", control < 3)
ipeds_filter(stabbr \%in\% c("KY","TN"))
ipeds_filter(grepl("community", instnm))

# more complex filters, such as those using an `|` (OR) operator, should be
# input as a string, either directly or as stored in an object
ipeds_filter("stabbr == 'KY' | stabbr == 'TN'")

filter_str <- "stabbr == 'KY' | stabbr == 'TN'"
ipeds_filter(filter_str)
}

}
