% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ipeds_get}
\alias{ipeds_get}
\title{Get IPEDS data.}
\usage{
ipeds_get(
  ipedscall,
  bind = TRUE,
  join = TRUE,
  lower_names = TRUE,
  include_filter_vars = TRUE
)
}
\arguments{
\item{ipedscall}{Current list of parameters carried forward from prior
functions in the chain (ignore)}

\item{bind}{Row bind all same name survey files (e.g., HD2022 and HD2023)}

\item{join}{Join different name survey files by UNITID and year}

\item{lower_names}{Lower column names in returned data}

\item{include_filter_vars}{Include filtering variables in output table}
}
\description{
This function gets IPEDS data by compiling and
converting all the previous piped output into a single URL string
that is used to get the data.
}
