% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ipeds_init}
\alias{ipeds_init}
\title{Initialize chained request.}
\usage{
ipeds_init(local_dir = NA, revised_files = TRUE)
}
\description{
This function initializes the data request. It should always be
the first in the series of piped functions.
}
\examples{
\dontrun{
ipeds_init()
}
}
