% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{ipeds_select}
\alias{ipeds_select}
\title{Select IPEDS data variables.}
\usage{
ipeds_select(ipedscall, ...)
}
\arguments{
\item{ipedscall}{Current list of parameters carried forward from prior
functions in the chain (ignore)}

\item{...}{Desired variable names separated by commas (not case sensitive)}
}
\description{
This function is used to select the variables returned in the final dataset.
}
\examples{
\dontrun{
ipeds_select(UNITID)
ipeds_select(UNITID, INSTNM)
ipeds_select(unitid, instnm)
}
}
