% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{ipeds_select}
\alias{ipeds_select}
\title{Select IPEDS data variables.}
\usage{
ipeds_select(ipedscall, ...)
}
\arguments{
\item{ipedscall}{Current list of parameters carried forward from prior
functions in the chain (ignore)}

\item{...}{Desired variable names separated by commas (not case sensitive)}
}
\description{
This function is used to select the variables returned in the final dataset.
By default, non-standard evaluation is assumed, meaning that you can input
variable names without quotation, separated by commas. If you set
\code{use_nse = FALSE} in \code{\link[=ipeds_init]{ipeds_init()}}, then you must use quotations around
column names or use an external object.
}
\examples{
\dontrun{
# with non-standard evalution
ipeds_select(UNITID)
ipeds_select(UNITID, INSTNM)
ipeds_select(unitid, instnm)

# without non-standard evaluation
ipeds_select("UNITID")
ipeds_select("unitid", "instnm")
ipeds_select(c("unitid", "instnm"))

vars <- c("unitid", "instnm")
ipeds_select(vars)
}
}
