% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{ipeds_select}
\alias{ipeds_select}
\title{Select IPEDS data variables.}
\usage{
ipeds_select(ipedscall, ...)
}
\arguments{
\item{ipedscall}{Current list of parameters carried forward from prior
functions in the chain (ignore)}

\item{...}{Desired variable names separated by commas (not case sensitive)}
}
\description{
Select the variables returned in the final dataset. By default, non-standard
evaluation is assumed, meaning that you can input variable names without
quotation, separated by commas. If \code{use_nse = FALSE} is set in
\code{\link[=ipeds_init]{ipeds_init()}}, then you must use standard evaluation, e.g., string variable
names or an external object which contains string variable names.
}
\details{
\code{\link[=ipeds_select]{ipeds_select()}} is a required part of the \verb{ipeds_*()} chain.
}
\examples{
\dontrun{
# with non-standard evalution
ipeds_select(UNITID)
ipeds_select(UNITID, INSTNM)
ipeds_select(unitid, instnm)

# with standard evaluation: ipeds_init(use_nse = FALSE)
ipeds_select("UNITID")
ipeds_select("unitid", "instnm")
ipeds_select(c("unitid", "instnm"))

vars <- c("unitid", "instnm")
ipeds_select(vars)
}
}
