% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripeds-pkg.R
\name{ripeds}
\alias{ripeds}
\title{An R data-package for NCES IPEDS survey datasets}
\description{
An R data-package to easily have complete survey files for quick analysis of IPEDS data.
}
\author{
Brock Tibert
}

